requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
eval_type_backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fsspec==2025.7.0
gradio==5.46.1
gradio_client==1.13.1
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.0
idna==3.10
invoke==2.2.0
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistralai==1.9.10
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
protobuf==6.32.1
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.4
rich==14.1.0
ruff==0.13.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
semantic-version==2.10.0
sentence-transformers==5.0.0
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.2.2
tqdm==4.67.1
transformers==4.53.2
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.36.0
websockets==15.0.1
